refactor: move routes into separate files, add package for config
